+++ ACP3/Core/Assets/Libraries.php
@@ -57,6 +57,10 @@ class Libraries
             'css' => 'bootstrap-datetimepicker.css',
             'js' => 'bootstrap-datetimepicker.min.js',
         ],
+        'font-awesome' => [
+            'enabled' => false,
+            'css' => 'font-awesome.css',
+        ],
     ];
     /**
      * @var EventDispatcherInterface
@@ -96,6 +100,10 @@ class Libraries
     {
         if (!isset($this->libraries[$identifier])) {
             $this->libraries[$identifier] = $options;
+        }
+
+        if (isset($options['enabled']) && $options['enabled'] === true) {
+            $this->enableLibraries($options['dependencies'] ?? []);
         }
 
         return $this;