+++ ACP3/Modules/ACP3/Gallery/Validation/GalleryFormValidation.php
@@ -57,6 +57,11 @@ class GalleryFormValidation extends Core\Validation\AbstractFormValidation
             $formData,
             ['path' => $this->uriAlias]
         );
+        $this->validator->dispatchValidationEvent(
+            'share.validation.validate_sharing_info',
+            $formData,
+            ['path' => $this->uriAlias]
+        );
 
         $this->validator->validate();
     }