+++ ACP3/Modules/ACP3/Share/Controller/Admin/Index/Index.php
@@ -1,0 +1,89 @@
+<?php
+
+/**
+ * Copyright (c) by the ACP3 Developers.
+ * See the LICENSE file at the top-level module directory for licensing details.
+ */
+
+namespace ACP3\Modules\ACP3\Share\Controller\Admin\Index;
+
+use ACP3\Core;
+use ACP3\Modules\ACP3\Share;
+use ACP3\Modules\ACP3\System\Installer\Schema;
+
+class Index extends Core\Controller\AbstractFrontendAction
+{
+    /**
+     * @var \ACP3\Modules\ACP3\Share\Model\Repository\DataGridRepository
+     */
+    protected $dataGridRepository;
+
+    /**
+     * Index constructor.
+     *
+     * @param \ACP3\Core\Controller\Context\FrontendContext                $context
+     * @param \ACP3\Modules\ACP3\Share\Model\Repository\DataGridRepository $dataGridRepository
+     */
+    public function __construct(
+        Core\Controller\Context\FrontendContext $context,
+        Share\Model\Repository\DataGridRepository $dataGridRepository
+    ) {
+        parent::__construct($context);
+
+        $this->dataGridRepository = $dataGridRepository;
+    }
+
+    /**
+     * @return array
+     */
+    public function execute()
+    {
+        /** @var Core\Helpers\DataGrid $dataGrid */
+        $dataGrid = $this->get('core.helpers.data_grid');
+        $dataGrid
+            ->setRepository($this->dataGridRepository)
+            ->setRecordsPerPage($this->resultsPerPage->getResultsPerPage(Schema::MODULE_NAME))
+            ->setIdentifier('#share-data-grid')
+            ->setResourcePathDelete('admin/share/index/delete')
+            ->setResourcePathEdit('admin/share/index/edit');
+
+        $this->addDataGridColumns($dataGrid);
+
+        return [
+            'grid' => $dataGrid->render(),
+            'show_mass_delete_button' => $dataGrid->countDbResults() > 0,
+        ];
+    }
+
+    /**
+     * @param Core\Helpers\DataGrid $dataGrid
+     */
+    protected function addDataGridColumns(Core\Helpers\DataGrid $dataGrid)
+    {
+        $dataGrid
+            ->addColumn([
+                'label' => $this->translator->t('share', 'active'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\ReplaceValueColumnRenderer::class,
+                'fields' => ['active'],
+                'custom' => [
+                    'search' => [0, 1],
+                    'replace' => [$this->translator->t('system', 'no'), $this->translator->t('system', 'yes')],
+                ],
+            ], 30)
+            ->addColumn([
+                'label' => $this->translator->t('share', 'uri'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\RouteColumnRenderer::class,
+                'fields' => ['uri'],
+                'default_sort' => true,
+                'custom' => [
+                    'path' => '%s',
+                ],
+            ], 20)
+            ->addColumn([
+                'label' => $this->translator->t('system', 'id'),
+                'type' => Core\Helpers\DataGrid\ColumnRenderer\IntegerColumnRenderer::class,
+                'fields' => ['id'],
+                'primary' => true,
+            ], 10);
+    }
+}