+++ ACP3/Modules/ACP3/Share/Model/Repository/DataGridRepository.php
@@ -1,0 +1,13 @@
+<?php
+
+/**
+ * Copyright (c) by the ACP3 Developers.
+ * See the LICENSE file at the top-level module directory for licensing details.
+ */
+
+namespace ACP3\Modules\ACP3\Share\Model\Repository;
+
+class DataGridRepository extends \ACP3\Core\Model\Repository\DataGridRepository
+{
+    const TABLE_NAME = ShareRepository::TABLE_NAME;
+}